fix build

+++ deploy.ps1
@@ -7,9 +7,12 @@ param(
 )
 
 
-Remove-Item -Recurse -Force "build"
-Remove-Item -Recurse -Force "dist"
+if (Test-Path -Path "build") {
+    Remove-Item -Recurse -Force "build"
+}
+if (Test-Path -Path "dist") {
+    Remove-Item -Recurse -Force "dist"
+}
 
-python setup.py bdist_wheel
-
-twine upload --repository-url $RepoUrl -u $User -p $Pwd dist/*
+& python setup.py bdist_wheel
+& twine upload --repository-url $RepoUrl -u $User -p $Pwd dist/*
